scripts/requirements.txt
requests
transformers
pillow
openai